Add test and benchmark for BlumPrime
This adds a simple test making sure that the output is actually a prime
number. Later, this will be expanded to make sure that the output is
safe (which implies Blum, as well).

This test isn't deterministic, and you probably want to do:

  go test -count=1

to avoid using cached results.

This also includes a benchmark for BlumPrime.