Define BytesWithDomain type
This is a useful wrapper allowing us to ascribe a specific domain to some
piece of data. This allows us to have domain separation for different byte
slices, without creating a wrapper type each time.

That being said, it probably would be a good idea to define wrapper types
for some of these, but I digress.