Extend WriterTo implementations with Domain
There are currently many types implementing WriterTo, and this extends
those types to also implement WriterToWithDomain, by adding a Domain() method.

The usual convention for these types is to simply use the name of the type
as the domain. The exception are Paillier related types. For these, I felt
that 'Paillier Ciphertext' was a better domain string than just 'Ciphertext'.

The important thing with domains is that they're unique (at least in
this library). Having a naming convention helps avoid this, but pretty names
are more of an aesthetic concern.