Always use domain separation in Hash.WriteAny
This means that any call to this function will not produce any conflicts.

To accomplish this, it adds its own domain separation for concrete types.
It also accepts the WriterTowithDomain interface, instead of just WriterTo,
that way it can use the domain separation that comes along.

By the way, this function would be a great candidate for generics, because you
could make it type safe. Right now, it's very scary mucking with this function,
because you have no idea how much other code might break. This despite
making sure you pass the right types to this function could be statically
checked.