Remove equality assertion betwen ZK protobufs
This is very fragile, and fails to actually check equality. This is
because of hidden fields in safenum.Nat which don't matter for equality,
but exist for optimization reasons.

My reasoning for removing these is that the proof verification should
fail if marshalling returned a different result anyways, so that should
be sufficient to check roundtripping errors.

We could perhaps amend this situation by implementing an Equal method on
protobufs ourselves, which is a possibility.