Improve extended OT monochrome check based on paper revision
This switches the monochrome check in the extended OT implementation to
use the revision of the paper, which corrects the security proof. The
approach is very different and much more robust.

One difference is that the generation of the random chi check factors is
done by hashing the current protocol context. The paper suggests adding
an extra round for the checking party to send a random seed to include
in this hash as well. Afaict, this is to make the security proof go
through easily, and I don't think this is strictly necessary.

The plumbing to add an extra round is annoying, so that can be done in a
subsequent commit if desired.